Fix alignment OOM, migrate index to 64-bit, build bwa-mem2-style 2N index

+++ tests/fm_index_large_ref.rs
@@ -173,7 +173,7 @@ fn test_f_column_correctness() {
         }
         // F[4] - 1 should be >= reference length
         assert!(
-            index.f_column[4] >= index.len as u32,
+            index.f_column[4] >= index.len as u64,
             "F[4]={} should be >= len={} for size {}",
             index.f_column[4],
             index.len,
